refactor: move settings classes to correct layer
