+++ src/ei8-Cortex-Subscriptions-Master.sln
@@ -23,8 +23,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Persistence", "Persistence"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SQLite", "main\Port.Adapter\IO\Persistence\SQLite\SQLite.csproj", "{81552235-BE3A-4801-B490-74DF1738F050}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Common", "..\..\cortex-subscriptions-common\src\main\Common.csproj", "{817DDCFA-56D1-4EBA-81F5-5FE9495A10B7}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Domain.Model", "main\Domain.Model\Domain.Model.csproj", "{7CD57BE5-09E7-48D3-AA49-4203AE02403A}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{A6EE74C5-CFF3-438A-935B-EE84E876137B}"
@@ -32,7 +30,9 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		.editorconfig = .editorconfig
 	EndProjectSection
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Http", "..\..\ei8-net-http\src\main\Http.csproj", "{41464C5D-D952-4724-BC1B-847D5A695F73}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Process", "Process", "{B1EF9935-3ABD-4F23-81B8-FDBFCB5B38DC}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Services", "main\Port.Adapter\IO\Process\Services\Services.csproj", "{5EED8224-F9DA-4149-BC75-F8C492A0B0F0}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -64,18 +64,14 @@ Global
 		{81552235-BE3A-4801-B490-74DF1738F050}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{81552235-BE3A-4801-B490-74DF1738F050}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{81552235-BE3A-4801-B490-74DF1738F050}.Release|Any CPU.Build.0 = Release|Any CPU
-		{817DDCFA-56D1-4EBA-81F5-5FE9495A10B7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{817DDCFA-56D1-4EBA-81F5-5FE9495A10B7}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{817DDCFA-56D1-4EBA-81F5-5FE9495A10B7}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{817DDCFA-56D1-4EBA-81F5-5FE9495A10B7}.Release|Any CPU.Build.0 = Release|Any CPU
 		{7CD57BE5-09E7-48D3-AA49-4203AE02403A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{7CD57BE5-09E7-48D3-AA49-4203AE02403A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{7CD57BE5-09E7-48D3-AA49-4203AE02403A}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{7CD57BE5-09E7-48D3-AA49-4203AE02403A}.Release|Any CPU.Build.0 = Release|Any CPU
-		{41464C5D-D952-4724-BC1B-847D5A695F73}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{41464C5D-D952-4724-BC1B-847D5A695F73}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{41464C5D-D952-4724-BC1B-847D5A695F73}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{41464C5D-D952-4724-BC1B-847D5A695F73}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5EED8224-F9DA-4149-BC75-F8C492A0B0F0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5EED8224-F9DA-4149-BC75-F8C492A0B0F0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5EED8224-F9DA-4149-BC75-F8C492A0B0F0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5EED8224-F9DA-4149-BC75-F8C492A0B0F0}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -86,6 +82,8 @@ Global
 		{978807E3-75B3-4665-BA42-D97194428421} = {24D6B38C-733D-4B1F-93C7-FBA2D13D3718}
 		{4F60AA9F-6CA3-4E9C-9540-A68AE34B9F0C} = {978807E3-75B3-4665-BA42-D97194428421}
 		{81552235-BE3A-4801-B490-74DF1738F050} = {4F60AA9F-6CA3-4E9C-9540-A68AE34B9F0C}
+		{B1EF9935-3ABD-4F23-81B8-FDBFCB5B38DC} = {978807E3-75B3-4665-BA42-D97194428421}
+		{5EED8224-F9DA-4149-BC75-F8C492A0B0F0} = {B1EF9935-3ABD-4F23-81B8-FDBFCB5B38DC}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {C5E7B24F-6EE6-48A2-A996-5075DDAF0A4B}
