+++ src/main/Application/Application.csproj
@@ -7,8 +7,10 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\..\..\cortex-subscriptions-common\src\main\Common.csproj" />
-    <ProjectReference Include="..\..\..\..\ei8-net-http\src\main\Http.csproj" />
+    <PackageReference Include="ei8.Cortex.Subscriptions.Common" Version="0.1.0" />
+  </ItemGroup>
+
+  <ItemGroup>
     <ProjectReference Include="..\Domain.Model\Domain.Model.csproj" />
   </ItemGroup>
