+++ src/main/Application/ISettingsService.cs
@@ -1,6 +1,6 @@
 ﻿using ei8.Net.Http.Notifications;
 
-namespace ei8.Cortex.Subscriptions.Domain.Model
+namespace ei8.Cortex.Subscriptions.Application
 {
     public interface ISettingsService
     {
