+++ src/main/Port.Adapter/IO/Process/Services/Services.csproj
@@ -1,0 +1,17 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netstandard2.1</TargetFramework>
+    <AssemblyName>ei8.Cortex.Subscriptions.Port.Adapter.IO.Process.$(MSBuildProjectName)</AssemblyName>
+    <RootNamespace>ei8.Cortex.Subscriptions.Port.Adapter.IO.Process$(MSBuildProjectName.Replace(" ", "_"))</RootNamespace>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.Extensions.Configuration.Abstractions" Version="6.0.0" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\..\..\Application\Application.csproj" />
+  </ItemGroup>
+
+</Project>
