+++ src/main/Port.Adapter/In/Api/BackgroundServices/AvatarPollingBackgroundService.cs
@@ -1,4 +1,5 @@
-﻿using ei8.Cortex.Subscriptions.Application.Interface.Service;
+﻿using ei8.Cortex.Subscriptions.Application;
+using ei8.Cortex.Subscriptions.Application.Interface.Service;
 using ei8.Cortex.Subscriptions.Domain.Model;
 
 namespace ei8.Cortex.Subscriptions.In.Api.BackgroundServices
